refactor: handback payload builder moves to lib/delegation/handback.js
Factory-injected transcript access and marker stripping keep the module
free of workspace globals and directly unit-testable. server.js binds
it once and keeps exporting the same names, so the existing unit tests
run unchanged against the extracted implementation.
