+++ server.js
@@ -19,6 +19,7 @@ const PKG_VERSION = require('./package.json').version;
 const searchLib = require('./search.js');
 const { resolvePermissionConvoId } = require('./permission-routing.js');
 const { resolveMarkers } = require('./lib/delegation/markers.js');
+const { createHandbackBuilder } = require('./lib/delegation/handback.js');
 
 const PORT = process.env.PORT || 3000;
 let ACTUAL_PORT = PORT; // Updated after server.listen() with the real listening port
@@ -327,44 +328,15 @@ function sanitizeSpecialistOutput(text) {
   return cleaned;
 }
 
-// The delegate's plain chat turns from the on-disk transcript, in order.
-// Boundary is a TIMESTAMP, not an index: appendTranscript splices entry 1 at
-// the 1000-entry soft cap, so a stored index silently drifts on long
-// conversations. Typed rows (e.g. 'routing') are bookkeeping and carry no
-// session content.
-function transcriptTurnsSince(convoId, agentId, sinceIso) {
-  const transcript = loadTranscript(convoId) || [];
-  return transcript
-    .filter(t => t.role === 'agent' && t.agent === agentId && !t.type)
-    .filter(t => !sinceIso || t.timestamp >= sinceIso)
-    .map(t => t.text);
-}
-
-// Build the payload a parent agent receives when a delegate hands back. The
-// single source of handback content: every substantive turn, not just the
-// last one.
-//
-// Incident (0.11.2): an analyst delivered a 6,665-char analysis in turn 1 and
-// a 106-char sign-off in turn 2; only the sign-off reached the lead, because
-// finalResponseText holds one turn by design (responseText resets per turn).
-// The lead refused to invent the analysis and the user pasted 6,050 chars by
-// hand while the full report sat in the transcript on disk.
-//
-// Prefers in-memory accumulated turns (fast path); falls back to the
-// transcript, which survives process death and is authoritative. Truncation
-// over the cap is LOUD: the parent is told what was omitted and where the
-// full output lives, never silently handed a fragment.
-function buildHandbackPayload(entry, convoId) {
-  const turns = (entry.deliveredTurns && entry.deliveredTurns.length)
-    ? entry.deliveredTurns
-    : transcriptTurnsSince(convoId, entry.agentId, entry.delegationStartedAt);
-  const cleaned = turns.map(t => stripRundockMarkers(t || '').trim()).filter(Boolean);
-  const joined = cleaned.join('\n\n');
-  if (joined.length <= SPECIALIST_OUTPUT_MAX_CHARS) return joined;
-  const omitted = joined.length - SPECIALIST_OUTPUT_MAX_CHARS;
-  return joined.substring(0, SPECIALIST_OUTPUT_MAX_CHARS)
-    + `\n\n[Handback truncated: ${omitted} of ${joined.length} characters omitted across ${cleaned.length} turn(s). Read the full output in .rundock/transcripts/${convoId}.json]`;
-}
+// Handback payload building lives in lib/delegation/handback.js; the factory
+// receives transcript access and marker stripping so the module stays free
+// of workspace globals. Function declarations hoist, so binding here at the
+// top of the file is safe even though loadTranscript is defined further down.
+const { buildHandbackPayload, transcriptTurnsSince } = createHandbackBuilder({
+  loadTranscript: (convoId) => loadTranscript(convoId),
+  stripMarkers: (text) => stripRundockMarkers(text),
+  maxChars: SPECIALIST_OUTPUT_MAX_CHARS,
+});
 
 // Session history: read Claude Code JSONL transcripts from disk
 function getSessionJsonlPath(sessionId) {
